+++ src/main/java/org/catalytic/sdk/entities/CredentialType.java
@@ -1,9 +1,17 @@
 package org.catalytic.sdk.entities;
 
-/**
- * Contains all the possible values for the type of credential
- */
-public class CredentialType {
-    public static final String USER = "user";
-    public static final String ACTIONWORKER = "actionWorker";
+public enum CredentialType {
+
+    USER("user"),
+    ACTIONWORKER("actionWorker");
+
+    private String value;
+
+    CredentialType(String value) {
+        this.value = value;
+    }
+
+    public String getValue() {
+        return value;
+    }
 }